edit distance updated
