+++ torchmetrics/functional/text/mer.py
@@ -17,27 +17,7 @@ from typing import List, Tuple, Union
 import torch
 from torch import Tensor, tensor
 
-
-def _edit_distance(prediction_tokens: List[str], reference_tokens: List[str]) -> int:
-    """Standard dynamic programming algorithm to compute the edit distance.
-
-    Args:
-        prediction_tokens: A tokenized predicted sentence
-        reference_tokens: A tokenized reference sentence
-
-    Returns:
-        Editing distance between the predicted sentence and the reference sentence
-    """
-    dp = [[0] * (len(reference_tokens) + 1) for _ in range(len(prediction_tokens) + 1)]
-    dp[:][0] = list(range(len(prediction_tokens) + 1))
-    dp[0][:] = list(range(len(reference_tokens) + 1))
-    for i in range(1, len(prediction_tokens) + 1):
-        for j in range(1, len(reference_tokens) + 1):
-            if prediction_tokens[i - 1] == reference_tokens[j - 1]:
-                dp[i][j] = dp[i - 1][j - 1]
-            else:
-                dp[i][j] = min(dp[i - 1][j], dp[i][j - 1], dp[i - 1][j - 1]) + 1
-    return dp[-1][-1]
+from torchmetrics.functional.text.helper import _edit_distance
 
 
 def _mer_update(
